backend/pyproject.toml
[tool.poetry]
name = "lab08"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "*"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
pydantic = {extras = ["email"], version = "^2.12.5"}
python-dotenv = "*"
python-jose = {extras = ["cryptography"], version = "*"}
bcrypt = "^4.0.0"
faker = "^24.0.0"
black = "^25.11.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
